Adding scramble_bot.py
Scramble with friends bot and the supporting words file. This bot is
used to play the Zynga game "Scramble with friends". It uses a simple
brute force algorithm implemented in a semi-efficient manner using
python's built in structures. This is a monkey runner script and the
touch co-ordinate values are hard coded for Nexus 7 in Portrait
orientation.